requirements/base.txt
django==5.0.6
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.4.0
django-filter==24.2
django-ratelimit==4.1.0
python-decouple==3.8
celery==5.3.6
redis==5.0.7
django-celery-beat==2.6.0
django-celery-results==2.5.1
django-storages==1.14.4
boto3==1.34.144
firebase-admin==6.5.0
requests==2.32.3
pillow==10.4.0